router: move config and configwatcher packages to internal/
Implement Phase 1 of newrouter migration by relocating shared infrastructure
packages out from under proxy/ so the new router does not depend on the
legacy proxy tree.

- Move proxy/config → internal/config (17 files)
- Move proxy/configwatcher → internal/watcher (2 files)
- Update import paths in 33 files across cmd/, proxy/, internal/, and llama-swap.go
- Fix pre-existing data race in internal/router/server.go (wg.Wait() was
  deferred after errors.Join read the results slice)
- Update docs/newrouter-todo.md to mark Phase 1 complete
