+++ internal/config/peer_test.go
