+++ internal/config/performance.go
