+++ internal/process/process_command.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/mostlygeek/llama-swap/internal/logmon"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 var ErrStartAborted = fmt.Errorf("aborted")
