+++ internal/router/matrix_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/mostlygeek/llama-swap/internal/logmon"
 	"github.com/mostlygeek/llama-swap/internal/process"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 // newTestMatrix builds a Matrix router from supplied processes, bypassing
