+++ internal/router/peer_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/mostlygeek/llama-swap/internal/logmon"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 var testLogger = logmon.NewWriter(os.Stdout)
