+++ proxy/matrix.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	"github.com/mostlygeek/llama-swap/internal/logmon"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 // MatrixSolver contains pure swap-decision logic with no Process dependencies.
