+++ proxy/process.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/mostlygeek/llama-swap/event"
 	"github.com/mostlygeek/llama-swap/internal/logmon"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 type ProcessState string
