+++ proxy/processgroup.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 
 	"github.com/mostlygeek/llama-swap/internal/logmon"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 )
 
 type ProcessGroup struct {
