+++ proxy/proxymanager_test.go
@@ -17,7 +17,7 @@ import (
 	"time"
 
 	"github.com/mostlygeek/llama-swap/event"
-	"github.com/mostlygeek/llama-swap/proxy/config"
+	"github.com/mostlygeek/llama-swap/internal/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/tidwall/gjson"
 )
